Assignment1/requirements.txt
kafka-python
requests
faust
pymongo
